+++ packages/sit-onyx/src/components/OnyxListItem/OnyxListItem.vue
@@ -1,11 +1,14 @@
 <script setup lang="ts">
-const props = defineProps<{ active?: boolean }>();
+import type { ListboxOption } from "../OnyxListbox/types";
+
+const props = defineProps<Pick<ListboxOption, "color"> & { active?: boolean }>();
 </script>
 <template>
   <li
     :class="{
       'onyx-list-item': true,
       'onyx-list-item--active': props.active,
+      'onyx-list-item--danger': props.color === 'danger',
     }"
   >
     <slot></slot>
@@ -13,11 +16,68 @@ const props = defineProps<{ active?: boolean }>();
 </template>
 <style lang="scss">
 @use "../../styles/mixins/layers";
-@use "../../styles/mixins/listitem";
 
 .onyx-list-item {
   @include layers.component() {
-    @include listitem.styles(&);
+    --onyx-list-item-color: var(--onyx-color-text-icons-neutral-intense);
+    --onyx-list-item-color-selected: var(--onyx-color-text-icons-primary-bold);
+    --onyx-list-item-background-hover: var(--onyx-color-base-primary-100);
+    --onyx-list-item-background-selected: var(--onyx-color-base-primary-200);
+
+    font-family: var(--onyx-font-family);
+    color: var(--onyx-list-item-color);
+    padding: var(--onyx-spacing-2xs) var(--onyx-spacing-sm);
+    background-color: var(--onyx-color-base-background-blank);
+    font-weight: 400;
+    font-size: 1rem;
+    line-height: 1.5rem;
+
+    border: none;
+    width: 100%;
+    outline: none;
+
+    display: flex;
+    align-items: center;
+    gap: var(--onyx-spacing-sm);
+
+    &:not([aria-disabled="true"]) {
+      cursor: pointer;
+
+      &:hover,
+      &.onyx-list-item--active {
+        background-color: var(--onyx-list-item-background-hover);
+      }
+
+      // single select
+      &[aria-selected="true"] {
+        background-color: var(--onyx-list-item-background-selected);
+
+        &:hover,
+        &.onyx-list-item--active {
+          color: var(--onyx-list-item-color-selected);
+        }
+      }
+
+      // multiselect
+      &[aria-checked="true"] {
+        &:hover,
+        &.onyx-list-item--active {
+          background-color: var(--onyx-list-item-background-selected);
+          color: var(--onyx-list-item-color-selected);
+        }
+      }
+    }
+
+    &[aria-disabled="true"] {
+      color: var(--onyx-color-text-icons-neutral-soft);
+    }
+
+    &--danger {
+      --onyx-list-item-color: var(--onyx-color-text-icons-danger-intense);
+      --onyx-list-item-color-selected: var(--onyx-color-text-icons-danger-intense);
+      --onyx-list-item-background-hover: var(--onyx-color-base-danger-100);
+      --onyx-list-item-background-selected: var(--onyx-color-base-danger-200);
+    }
   }
 }
 </style>