+++ packages/sit-onyx/src/components/OnyxListboxOption/OnyxListboxOption.ct.tsx
@@ -1,9 +1,21 @@
 import { test } from "../../playwright/a11y";
+import { DENSITIES } from "../../composables/density";
 import { executeMatrixScreenshotTest } from "../../playwright/screenshots";
 import OnyxListboxOption from "./OnyxListboxOption.vue";
 
+const disabledRules = [
+  // aria-required-parent is ignored here because this component is only a single option which is internally always
+  // used together with a parent so we disable the failing rule here
+  "aria-required-parent",
+  // TODO: color-contrast: remove when contrast issues are fixed in https://github.com/SchwarzIT/onyx/issues/410
+  "color-contrast",
+  // TODO: as part of https://github.com/SchwarzIT/onyx/issues/1026,
+  // the following disabled rule should be removed.
+  "nested-interactive",
+];
+
 test.describe("Single select screenshot tests", () => {
-  for (const state of ["default", "disabled"] as const) {
+  for (const state of ["default", "disabled", "danger"] as const) {
     executeMatrixScreenshotTest({
       name: `Listbox option (${state})`,
       columns: ["default", "selected"],
@@ -18,6 +30,7 @@ test.describe("Single select screenshot tests", () => {
           aria-selected={column === "selected"}
           active={row === "focus-visible"}
           aria-disabled={state === "disabled"}
+          color={state === "danger" ? "danger" : undefined}
           {...{ role: "option" }}
         >
           Test label
@@ -36,22 +49,13 @@ test.describe("Multiselect Screenshot tests", () => {
       name: `Multiselect Listbox option (${state})`,
       columns: ["default", "checked", "indeterminate"],
       rows: ["default", "hover", "focus-visible"],
-      // aria-required-parent is ignored here because this component is only a single option which is internally always
-      // used together with a parent so we disable the failing rule here
-      disabledAccessibilityRules: [
-        "aria-required-parent",
-        // TODO: color-contrast: remove when contrast issues are fixed in https://github.com/SchwarzIT/onyx/issues/410
-        "color-contrast",
-        // TODO: as part of https://github.com/SchwarzIT/onyx/issues/1026,
-        // the following disabled rule should be removed.
-        "nested-interactive",
-      ],
+      disabledAccessibilityRules: disabledRules,
       component: (column, row) => (
         <OnyxListboxOption
           aria-label="Label"
           aria-checked={column === "checked"}
-          active={row === "focus-visible"}
           aria-disabled={state === "disabled"}
+          active={row === "focus-visible"}
           multiple={true}
           indeterminate={column === "indeterminate"}
           {...{ role: "option" }}
@@ -64,4 +68,26 @@ test.describe("Multiselect Screenshot tests", () => {
       },
     });
   }
+});
+
+test.describe("Density Screenshot tests", () => {
+  executeMatrixScreenshotTest({
+    name: "Listbox option (densities)",
+    columns: DENSITIES,
+    rows: ["default", "multiselect"],
+    disabledAccessibilityRules: disabledRules,
+    component: (column, row) => (
+      <OnyxListboxOption
+        aria-label="Label"
+        multiple={row === "multiselect"}
+        density={column}
+        {...{ role: "option" }}
+      >
+        Test label
+      </OnyxListboxOption>
+    ),
+    beforeScreenshot: async (component) => {
+      await component.hover();
+    },
+  });
 });