+++ packages/sit-onyx/src/components/OnyxListboxOption/OnyxListboxOption.vue
@@ -1,7 +1,16 @@
 <script lang="ts" setup>
+import { useDensity } from "../../composables/density";
+import OnyxIcon from "../OnyxIcon/OnyxIcon.vue";
+import OnyxListItem from "../OnyxListItem/OnyxListItem.vue";
 import type { OnyxListboxOptionProps } from "./types";
 
-const props = defineProps<OnyxListboxOptionProps>();
+const props = withDefaults(defineProps<OnyxListboxOptionProps>(), {
+  active: false,
+  multiple: false,
+  color: "primary",
+});
+
+const { densityClass } = useDensity(props);
 
 defineSlots<{
   /**
@@ -12,7 +21,11 @@ defineSlots<{
 </script>
 
 <template>
-  <li class="onyx-listbox-option" :class="{ 'onyx-listbox-option--active': props.active }">
+  <OnyxListItem
+    :class="['onyx-listbox-option', densityClass]"
+    :active="props.active"
+    :color="props.color"
+  >
     <input
       v-if="props.multiple"
       :checked="!!$attrs['aria-checked']"
@@ -25,10 +38,12 @@ defineSlots<{
       type="checkbox"
     />
 
+    <OnyxIcon v-if="props.icon" :icon="props.icon" />
+
     <span class="onyx-truncation-ellipsis">
       <slot></slot>
     </span>
-  </li>
+  </OnyxListItem>
 </template>
 
 <style lang="scss">
@@ -40,8 +55,6 @@ defineSlots<{
   @include checkbox.variables();
 
   @include layers.component() {
-    @include listitem.styles(&);
-
     &__checkbox {
       @include checkbox.styles();
       // prevent the checkbox to get squished by a long label