+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,20 @@
 # alpha-test-app
 
+## 0.1.0-alpha.90
+
+### Patch Changes
+
+- Updated dependencies [c55a599]
+  - sit-onyx@1.0.0-alpha.93
+
+## 0.1.0-alpha.89
+
+### Patch Changes
+
+- Updated dependencies [2b48da9]
+- Updated dependencies [2b48da9]
+  - sit-onyx@1.0.0-alpha.92
+
 ## 0.1.0-alpha.88
 
 ### Patch Changes