+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,20 @@
 # playground
 
+## 0.0.1-alpha.32
+
+### Patch Changes
+
+- Updated dependencies [c55a599]
+  - sit-onyx@1.0.0-alpha.93
+
+## 0.0.1-alpha.31
+
+### Patch Changes
+
+- Updated dependencies [2b48da9]
+- Updated dependencies [2b48da9]
+  - sit-onyx@1.0.0-alpha.92
+
 ## 0.0.1-alpha.30
 
 ### Patch Changes